requirements.txt
# =============================================================================
# DBAI Python-Abhängigkeiten — Multi-Arch (x86_64 + arm64)
# =============================================================================

# ─── Datenbank ────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# ─── Hardware-Monitoring ──────────────────────────────────────────
psutil>=5.9.0

# ─── LLM (llama.cpp — kompiliert automatisch für x86/ARM) ────────
# llama-cpp-python erkennt automatisch:
#   x86_64: AVX, AVX2, CUDA (wenn NVIDIA vorhanden)
#   arm64:  NEON, Apple Neural Engine (wenn verfügbar)
llama-cpp-python>=0.2.0

# ─── Konfiguration ───────────────────────────────────────────────
toml>=0.10.2

# ─── GPU-Management ──────────────────────────────────────────────
# NVIDIA (nur x86_64 — auf ARM wird es ignoriert)
pynvml>=12.0.0; platform_machine=="x86_64"

# ─── Web-Server (Neural Bridge) ─────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
aiohttp>=3.9.0
aiofiles>=23.0.0
httpx>=0.27.0

# ─── Authentifizierung ──────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.1.0

# ─── Kryptografie ───────────────────────────────────────────────
cryptography>=42.0.0

# ─── Tests ───────────────────────────────────────────────────────
pytest>=7.0.0
pytest-cov>=4.0.0

# ─── Security-Immunsystem ───────────────────────────────────────
python-nmap>=0.7.1          # Nmap-Integration (Kali-Sidecar)
schedule>=1.2.0             # Cron-ähnlicher Task-Scheduler
scapy>=2.5.0                # Netzwerk-Paket-Analyse
paramiko>=3.4.0             # SSH-Security-Checks
pip-audit>=2.7.0            # Python-Dependency-CVE-Check

# ─── Plattform-Hinweise ─────────────────────────────────────────
# x86_64:  CUDA/AVX beschleunigt LLM-Inferenz stark
# arm64:   ARM NEON reicht für kleine Modelle (7B-13B)
# Beide:   PostgreSQL + pgvector identisch
